+++ src/main/java/faang/school/analytics/handler/GlobalExceptionHandler.java
@@ -1,11 +1,10 @@
 package faang.school.analytics.handler;
 
+import faang.school.analytics.exception.ErrorResponse;
+import faang.school.analytics.exception.NotFoundException;
 import faang.school.analytics.exception.ValidationException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.convert.ConversionFailedException;
-import faang.school.analytics.exception.ErrorResponse;
-import faang.school.analytics.exception.NotFoundException;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -27,6 +26,8 @@ public class GlobalExceptionHandler {
                 .message(e.getMessage())
                 .causeMessage(e.getCause().getMessage())
                 .build();
+    }
+
     @ExceptionHandler(NotFoundException.class)
     @ResponseStatus(HttpStatus.NOT_FOUND)
     public ErrorResponse handleNotFoundException(NotFoundException e) {
